Combine student data and SGP results

combineSGP(sgp_object,
        state=NULL,
        years=NULL,
        content_areas=NULL,
        sgp.percentiles=TRUE,
        sgp.percentiles.baseline=TRUE,
        sgp.projections=TRUE,
        sgp.projections.baseline=TRUE,
        sgp.projections.lagged=TRUE,
        sgp.projections.lagged.baseline=TRUE,
        sgp.target.scale.scores=FALSE,
        sgp.target.scale.scores.only=FALSE,
    sgp.target.scale.scores.merge=FALSE,
        sgp.target.content_areas=NULL,
        max.sgp.target.years.forward=3,
        update.all.years=FALSE,
        sgp.config=NULL,
        sgp.percentiles.equated=NULL,
        SGPt=NULL,
  fix.duplicates=NULL,
        parallel.config=NULL)

Arguments

sgp_object
An object of class SGP containing slots @Data (from prepareSGP) and @SGP (from analyzeSGP).

state
Acronym for which state is to be used for the lagged projections and growth to standard analyses. Function will try to guess state name from passed sgp_object is missing.

years
A vector of years indicating years of data to merge with @Data. If missing, merge will use all available years of student growth percentile results.

content_areas
A vector of content areas indicating content areas of student growth percentile data to merge with @Data. If missing, merge will use all available content areas of student growth percentile results.

sgp.percentiles
A Boolean variable indicating whether to combine student growth percentiles.

sgp.percentiles.baseline
A Boolean variable indicating whether to combine baseline student growth percentiles.

sgp.projections
A Boolean variable indicating whether to combine current year student growth projections and calculate catch-up/keep-up and move-up/stay-up values.

sgp.projections.baseline
A Boolean variable indicating whether to combine current year baseline student growth projections and calculate catch-up/keep-up and move-up/stay-up values.

sgp.projections.lagged
A Boolean variable indicating whether to combine lagged student growth projections and calculate catch-up/keep-up and move-up/stay-up values.

sgp.projections.lagged.baseline
A Boolean variable indicating whether to combine lagged baseline student growth projections and calculate catch-up/keep-up move-up/stay-up values.

sgp.target.scale.scores
A Boolean variable indicating whether scale scores from calculated SGP targets should be calculated.

sgp.target.scale.scores.only
A Boolean variable indicating whether combineSGP should skip merging and only calculate scale scores from calculated SGP targets. Default is FALSE.

sgp.target.scale.scores.merge
An argument (defaults to FALSE) indicating whether & how target scale scores should be merged with the LONG data in the @Data slot.

sgp.target.content_areas
A Boolean variable indicating whether content area associated with SGP targets should be calculated.

max.sgp.target.years.forward
A integer indicating the number of years forward from the lagged (last year's) score to project forward for growth to standard calculations. Default is 3 years from the present, 4 years from the lagged year, which is the standard in most growth to standard calculations used by state departments of education.

update.all.years
A Boolean argument defaulting to FALSE indicating whether combineSGP should delete previously merged variables calculated in analyzeSGP and re-merge all available data.

sgp.config
Argument (defaults to NULL) passed utilized only for target scale score calculation. If targets for end of course tests are required, user must specify configurations directly. See code from testSGP number 3 for an example.

sgp.percentiles.equated
Boolean variable indicating whether equated percentiles are being merged (defaults to NULL which is converted to FALSE unless otherwise explicitly declared. If scale score targets are being calculated, linkages will be passed to scale score targets for calculation.

SGPt
Argument used to pass SGPt configuration to getTargetScaleScore for studentGrowthProjections calculation of scale score targets.

fix.duplicates
Argument to control how duplicate records based upon the key of VALID_CASE, CONTENT_AREA, YEAR, and ID are dealt with. If set to 'KEEP.ALL', the function tries to fix the duplicate individual records. If present and a '_DUP_***' suffix has been added to the duplicate ID before running studentGrowthPercentiles, then the @Data slot will be extended as necessary to accomodate additional student records and SGP results.

parallel.config
Parallel configuration only used when 'sgp.target.scale.scores' is set to TRUE. Default is NULL consistent with no targets being calculated. To utilize parallel processing in the calculation of SGP scale score targets user must specify a list designating a backend (e.g, BACKEND='PARALLEL') and a number of workers (e.g., 'WORKERS=list(SGP_SCALE_SCORE_TARGETS=4)').

Value

Function returns a list containing the input long data set in the @Data slot as a data.table keyed using VALID_CASE, CONTENT_AREA, YEAR, ID merged with student growth percentiles and/or straight/lagged projection targets and catch-up/keep-up and move-up/stay-up status with, if requested, the scale scores associated with such targets.

See also

prepareSGP, analyzeSGP

Examples

not_run({
  ## combineSGP is Step 3 of 5 of abcSGP
  Demonstration_SGP <- sgpData_LONG
  Demonstration_SGP <- prepareSGP(Demonstration_SGP)
  Demonstration_SGP <- analyzeSGP(Demonstration_SGP)
  Demonstration_SGP <- combineSGP(Demonstration_SGP)
})