Identify potential course progressions for SGP analyses

courseProgressionSGP(
        sgp_object,
        lag.direction=c("FORWARD", "BACKWARD"),
        year)

Arguments

sgp_object
Either a panel data set in long form or an object of class SGP. See embedded sgpData_LONG data set for an exemplar.

lag.direction
Character string indicating whether the progressions should be produced prospectively or retrospectively.

year
Character indicating the value of YEAR that is the focus of analysis.

Value

Function returns a nested list class object. The final node of each nested list is a data.table which summarizes the number of students with a particular GRADE by CONTENT_AREA by YEAR course progression

See also

sgpData_LONG

Examples

not_run({
  ## Run courseProgressionSGP on the subset of the long data that contains
  ## ONLY mathematics related records (would realistically also contain EOCT math courses)
  Math_Data <- subset(SGPdata::sgpData_LONG, CONTENT_AREA == "MATHEMATICS")
  Math_Progressions <- courseProgressionSGP(Math_Data, lag.direction= "BACKWARD", year="2015_2016")

  ##  Examine results for Adcademic Year 2015-2016, 5th grade Mathematics.
  Math_Progressions[['BACKWARD']][['2015_2016']][['MATHEMATICS.05']]
  Math_Progressions[['BACKWARD']][['2015_2016']][['MATHEMATICS.05']][COUNT>100]
})