Wrapper function associated with SGP analyses for Renaissance Learning Incorporated (RLI) interim STAR assessments.

rliSGP is a wrapper function design to expedite SGP analyses on Renaissance Learning Incorporated (RLI) interim STAR assessments.

rliSGP(
        sgp_object,
    additional.data=NULL,
        state=NULL,
        content_areas=c("MATHEMATICS", "READING", "EARLY_LITERACY"),
    testing.window,
    eow.or.update="UPDATE",
        update.save.shell.only=FALSE,
    configuration.year,
        sgp.percentiles.baseline=TRUE,
        sgp.projections.baseline=TRUE,
        sgp.projections.lagged.baseline=FALSE,
        sgp.target.scale.scores=TRUE,
    update.ids=NULL,
        SGPt=TRUE,
        simulate.sgps=FALSE,
        save.intermediate.results=FALSE,
        coefficient.matrices=NULL,
        goodness.of.fit.print=FALSE,
        return.updated.shell=FALSE,
        fix.duplicates="KEEP.ALL",
        eow.calculate.sgps=FALSE,
        score.type="RASCH",
        cutscore.file.name="Cutscores.csv",
    parallel.config=NULL)

Arguments

sgp_object
An object of class SGP or a data.frame from which an SGP object can be produced. If data.frame is supplied, the current window's data will be created from the last window present in the data.frame.

additional.data
Current window's data for UPDATE or EOW (end of window) processing. Can be NULL if new data all fed into argument 'sgp_object' as data.frame.

state
Abbreviation for STATE/ORGANIZATION being analyzed. Currently only available for 'RLI' or 'RLI_UK'.

content_areas
Character vector indicating which content areas rliSGP should be run for ('MATHEMATICS', 'READING', and/or 'EARLY_LITERACY'). Default is all.

testing.window
Argument indicating testing window being analyzed. Either 'FALL', 'WINTER', 'SPRING' indicating what testing window's data is being analyzed.

eow.or.update
Argument indicating whether end-of-window or within-window updates are being done. Either 'EOW' (end of window) or 'UPDATE' (within window).

update.save.shell.only
Boolean argument indicating whether End of Window processing should only produce/save RLI_SGP_UPDATE_SHELL and bypass coefficient matrix production.

configuration.year
Year associated with SGP analyses (configuration year) to be used.

sgp.percentiles.baseline
Boolean variable passed to updateSGP to control whether baseline student growth percentiles are calculated. Default is TRUE.

sgp.projections.baseline
Boolean variable passed to updateSGP to control whether baseline student growth projections are calculated. Default is TRUE.

sgp.projections.lagged.baseline
Boolean variable passed to updateSGP to control whether lagged baseline student growth projections are calculated. Default is FALSE.

sgp.target.scale.scores
Boolean variable passed to updateSGP to control whether scale score targets are calculated. Default is TRUE.

update.ids
Argument that allows ID to be updated based upon a supplied two column data frame where the first column represents the ORIGINAL/OLD ID and the second column the NEW ID that is to replace it. NOTE that ID needn't be replaced in either 'sgp_object' nor 'additional.data' that are provided as part of the 'rliSGP' call. Substitution will occur on entire data set before analysis. Default is NULL, no ID replacement is performed.

SGPt
Argument indicating whether to perform time dependent SGP (SGPt) analyses. Current default is TRUE based upon transition to SGPt in Summer 2015.

simulate.sgps
A Boolean argument indicating whether to calculate SGP standard errors based upon CSEMs provided by either an embedded CSEM-scale score lookup or an embedded variable. Default is FALSE.

save.intermediate.results
Boolean variable passed to updateSGP indicating whether to save intermediate results from prepareSGP, analyzeSGP, and combineSGP as part of the rliSGP analyses.

coefficient.matrices
List of baseline matrices to be used for analyses (overrides what exists in SGPstateData). Defaults to NULL (use those in SGPstateData).

goodness.of.fit.print
Boolean argument passed to updateSGP indicating whether to print goodness of fit plots associated with outcome. Default is FALSE.

return.updated.shell
Boolean argument (defaults to FALSE) indicating whether to return sgp_object as part of call to rliSGP.

fix.duplicates
Character argument (defaults to 'KEEP.ALL') indicating how to deal with duplicates submitted as part of data for analysis. Setting 'fix.duplicates' to null sets no modification of duplicates.

eow.calculate.sgps
Boolean argument passed to updateSGP further passed to analyzeSGP indicating whether student growth percentiles are calculated as part of EOW analyses. Default is FALSE which results in just coefficient matrices being calculated.

score.type
Either RASCH (default) or STAR indicating what type of score is being supplied for analysis. The selection ensures the correct cutscores, matrices, and knots/boundaries are used for analyses.

cutscore.file.name
Cutscore file name and path for input and creation of RLI Cutscores. Default is Cutscores.csv.

parallel.config
Argument passed to abcSGP or updateSGP to activate parallel processing with an appropriate backend.

Value

Saves relevant results to disc for use in subsequent analyses.

Examples

not_run({
  ####### For Within Window processing

  testSGP('RLI')

  ####### For End-of-Window processing

})