Function for producing goodness of fit plots using existing SGP object

gofSGP creates goodness-of-fit plots in either PDF or PNG for showing SGP distribution by prior achievement level and prior scale score decile. These plots expand upon the plots currently produced with the studentGrowthPercentiles function.

gofSGP(
        sgp_object,
        state=NULL,
        years=NULL,
        content_areas=NULL,
        content_areas_prior=NULL,
        grades=NULL,
        ceiling.floor=TRUE,
        use.sgp="SGP",
        output.format="PDF",
        color.scale="reds.and.blues")

Arguments

sgp_object
The SGP object from which the goodness-of-fit data will be used.

state
The 'state' for the sgp_object. Derive from sgp_object name if not explicitly supplied.

years
The years that goodness-of-fit plots are requested. Default is to use all years available.

content_areas
The content area(s) that goodness-of-fit plots are requested. Default is to use all content areas available.

content_areas_prior
The content area(s) of the prior year which growth by achievement level is being produced that goodness-of-fit plots are requested. Default is to use all content areas available.

grades
The grade(s) that goodness-of-fit plots are requested. Default is to use all grade available.

ceiling.floor
Boolean variable to explictly control display of ceiling/floor portion of goodness of fit plot.

use.sgp
Character vectors (defaults to 'SGP') indicating what student growth percentile variable to calculate goodness-of-fit plots for.

output.format
Character vectors (defaults to 'PDF') indicating what driver to use to output results. Options currently include 'PDF', 'PNG' and 'SVG'. 'SVG' is currently experimental.

color.scale
Character vectors (defaults to 'red') indicating what color palette to use for creating percentile distribution table.

Value

Returns output ('PDF' , 'PNG' or 'SVG') associated with goodness-of-fit analyses.

See also

studentGrowthPercentiles

Examples

not_run({
  Demonstration_SGP <- abcSGP(sgpData_LONG)
  gofSGP(Demonstration_SGP)
})