Class "splineMatrix"

Objects from the Class

Objects can be created by calls of the form new("splineMatrix", ...), but this is not encouraged. Previously produced coefficient matrices MUST be bound to the IDENTICAL knots and boundaries used to create them. Use the function studentGrowthPercentiles instead.

Slots

.Data:

A coefficient matrix derived from studentGrowthPercentiles.

Knots:

A list(s) of numeric values used as the knots to generate the B-spline basis matrix in studentGrowthPercentiles.

Boundaries:

A list(s) of numeric values used as the Boundary.knotsto generate the B-spline basis matrix in studentGrowthPercentiles.

Content_Areas:

A list of time dependent content area names included in the data used to produce the coefficient matrix.

Grade_Progression:

A list of the time dependent grades included in the data used to produce matrices.

Time:

A list of the Times (e.g., years) measurements occurred included in the data used to produce matrices.

Time_Lags:

A list of the time lags/differences between Time (e.g., years) included in the data used to produce matrices.

Version:

A list including the version of the SGP package used to construct the splineMatrix object and the date the object was created.

Details

splineMatrix:

This class contains the S3 matrix class, inheriting its methods. The slot Knots should be one or more lists of numeric vector(s) used in the internal call to bs, which generates a B-spline basis matrix from student scores. There are typically with 4 values for the knots. Similarly, Boundaries are used in bs for the Boundary.knots argument. This is always two values which are at or slightly beyond the lowest and highest observed student scores. Content_Areas and Grade_Progression provide information about the data (sub)set used to produce the matrix.

See also

studentGrowthPercentiles